requirements.txt
# CampusIQ - Python Dependencies
# Production-ready multi-tenant college management system

# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
oracledb==2.0.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
PyJWT==2.8.0
python-jose==3.3.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
requests-oauthlib==1.3.1

# API & Validation
pydantic==2.5.3
marshmallow==3.20.1
flask-cors==4.0.0

# Rate Limiting & Caching
Flask-Limiter==3.5.0
redis==5.0.1
cachetools==5.3.2

# Security
bcrypt==4.1.2
cryptography==41.0.7
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Natural Language Processing (for QnA)
spacy==3.7.2
nltk==3.8.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
uuid==1.30

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-asyncio==0.23.2
factory-boy==3.3.0
faker==22.0.0
httpx==0.26.0

# Code Quality
flake8==6.1.0
black==23.12.1
isort==5.13.2
mypy==1.8.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Logging & Monitoring
structlog==24.1.0
sentry-sdk==1.39.1

# Documentation
flask-swagger-ui==4.11.1
flasgger==0.9.7.1

# AI & LLM
google-generativeai==0.3.2
